requirements.txt
fastapi==0.70.0
uvicorn==0.12
pydantic[dotenv]==1.8.2
requests>=2.26.0
python-keycloak~=0.25
python-dateutil==2.8.2
itsdangerous~=2.0
python-consul==1.1.0
aiohttp==3.8.3